+++ src/app/locales/en/components/archive-browser/-archive-state.js
@@ -5,6 +5,6 @@ export default {
   preserved: 'Preserved',
   verification_failed: 'Verification&nbsp;failed',
   failed: 'Failed',
-  purging: 'Purging',
+  deleting: 'Deleting',
   unknown: 'Unknown',
 };