+++ src/app/locales/en/components/archives-delete.js
@@ -1,10 +1,10 @@
-import itemsInfo from './archives-purge/items-info';
+import itemsInfo from './archives-delete/items-info';
 
 export default {
   header: {
     headerText: {
-      single: 'Purge archive',
-      multi: 'Purge archives',
+      single: 'Delete archive',
+      multi: 'Delete archives',
     },
   },
   body: {
@@ -16,9 +16,9 @@ export default {
       selectedCount: '{{count}} selected archives',
     },
     unknownDataset: '<em>unknown</em>',
-    aboutToPurge: {
-      detailed: 'You are about to purge following {{archivesText}} created from <strong>{{datasetName}}</strong> dataset:',
-      simpleMulti: 'You are about to purge <strong>{{count}}</strong> archives created from <strong>{{datasetName}}</strong> dataset.',
+    aboutToDelete: {
+      detailed: 'You are about to delete following {{archivesText}} created from <strong>{{datasetName}}</strong> dataset:',
+      simpleMulti: 'You are about to delete <strong>{{count}}</strong> archives created from <strong>{{datasetName}}</strong> dataset.',
     },
     irreversibleWarning: 'This operation is irreversible. All the data in {{archivesText}} will be lost.',
     nestedSharesInfo: 'In case any nested file or directory is shared, this operation will make the share detached.',
@@ -31,11 +31,11 @@ export default {
   },
   footer: {
     cancel: 'Cancel',
-    purgeSubmit: {
-      single: 'Purge archive',
-      multi: 'Purge {{count}} archives',
+    deleteSubmit: {
+      single: 'Delete archive',
+      multi: 'Delete {{count}} archives',
     },
   },
-  purgingArchives: 'purging some archive',
+  deletingArchives: 'deleting some archive',
   itemsInfo,
 };