+++ src/app/locales/en/translations.js
@@ -30,7 +30,7 @@ import datasetBrowser from './components/dataset-browser';
 import archiveBrowser from './components/archive-browser';
 import archiveProperties from './components/archive-properties';
 import archiveCreate from './components/archive-create';
-import archivesPurge from './components/archives-purge';
+import archivesDelete from './components/archives-delete';
 import itemsSelectBrowser from './components/items-select-browser';
 import datasetProtection from './components/dataset-protection';
 import filePath from './components/file-path';
@@ -87,7 +87,7 @@ const translations = {
     archiveBrowser,
     archiveProperties,
     archiveCreate,
-    archivesPurge,
+    archivesDelete,
     itemsSelectBrowser,
     datasetProtection,
     filePath,